--- react/styles/spreadsheet.less
@@ -1,9 +1,0 @@
-// make grid scrollbars invisible
-// https://stackoverflow.com/a/38994837
-.spreadsheet-grid.no-scrollbar {
-    -ms-overflow-style: none; /* Internet Explorer 10+ */
-    scrollbar-width: none; /* Firefox */
-}
-.spreadsheet-grid.no-scrollbar::-webkit-scrollbar {
-    display: none; /* Safari and Chrome */
-}
